+++ apps/brownfield-example-shared-tests/src/global.d.ts
@@ -1,0 +1,7 @@
+/* eslint-disable no-var */
+declare global {
+  var __brownieExampleCounter: number;
+  var __resetBrownieExampleStore: () => void;
+}
+
+export {};