+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @callstack/brownfield-cli
 
+## 3.7.0
+
+### Minor Changes
+
+- [#313](https://github.com/callstack/react-native-brownfield/pull/313) [`c153378`](https://github.com/callstack/react-native-brownfield/commit/c1533783c0a93372b3c12f08c1428766c0405226) Thanks [@alpharius-ck](https://github.com/alpharius-ck)! - Jest unit tests
+
 ## 3.6.1
 
 ### Patch Changes